LayoutTests/imported/w3c/web-platform-tests/selection/anonymous/details-ancestor-expected.txt


FAIL Selecting the default summary of <details> should report a DOM-visible ancestor assert_equals: expected "HTMLDetailsElement" but got "HTMLBodyElement"
